Moved root and auth components
chore: moved auth files

chore: moved root components
