+++ src/app/(root)/Hero.tsx
