+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Hero from '@/components/Home/Hero';
+import Hero from '@/app/(root)/Hero';
 
 export default function Home() {
   return (
